+++ server/start.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/EscanBE/evermint/v12/constants"
+	"github.com/EscanBE/evermint/v12/indexer"
 	rpcclient "github.com/cometbft/cometbft/rpc/client"
 	"io"
 	"net"
@@ -48,11 +49,9 @@ import (
 	pruningtypes "github.com/cosmos/cosmos-sdk/store/pruning/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/EscanBE/evermint/v12/indexer"
 	ethdebug "github.com/EscanBE/evermint/v12/rpc/namespaces/ethereum/debug"
 	"github.com/EscanBE/evermint/v12/server/config"
 	srvflags "github.com/EscanBE/evermint/v12/server/flags"
-	evertypes "github.com/EscanBE/evermint/v12/types"
 )
 
 // DBOpener is a function to open `application.db`, potentially with customized options.
@@ -195,7 +194,6 @@ which accepts a path for the resulting pprof file.
 	cmd.Flags().Int32(srvflags.JSONRPCLogsCap, config.DefaultLogsCap, "Sets the max number of results can be returned from single `eth_getLogs` query")
 	cmd.Flags().Int32(srvflags.JSONRPCBlockRangeCap, config.DefaultBlockRangeCap, "Sets the max block range allowed for `eth_getLogs` query")
 	cmd.Flags().Int(srvflags.JSONRPCMaxOpenConnections, config.DefaultMaxOpenConnections, "Sets the maximum number of simultaneous connections for the server listener") //nolint:lll
-	cmd.Flags().Bool(srvflags.JSONRPCEnableIndexer, false, "Enable the custom tx indexer for json-rpc")
 	cmd.Flags().Bool(srvflags.JSONRPCEnableMetrics, false, "Define if EVM rpc metrics server should be enabled")
 
 	cmd.Flags().String(srvflags.EVMTracer, config.DefaultEVMTracer, "the EVM tracer type to collect execution traces from the EVM transaction execution (json|struct|access_list|markdown)") //nolint:lll
@@ -354,7 +352,6 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 	if gRPCOnly {
 		logger.Info("starting node in query only mode; Tendermint is disabled")
 		config.GRPC.Enable = true
-		config.JSONRPC.EnableIndexer = false
 	} else {
 		logger.Info("starting node with ABCI Tendermint in-process")
 
@@ -388,7 +385,7 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 	// Add the tx service to the gRPC router. We only need to register this
 	// service if API or gRPC or JSONRPC is enabled, and avoid doing so in the general
 	// case, because it spawns a new local tendermint RPC client.
-	if (config.API.Enable || config.GRPC.Enable || config.JSONRPC.Enable || config.JSONRPC.EnableIndexer) && tmNode != nil {
+	if (config.API.Enable || config.GRPC.Enable || config.JSONRPC.Enable) && tmNode != nil {
 		clientCtx = clientCtx.WithClient(local.New(tmNode))
 
 		app.RegisterTxService(clientCtx)
@@ -405,33 +402,6 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 	// Flag not added in config to avoid user enabling in config without passing in CLI
 	if config.JSONRPC.Enable && ctx.Viper.GetBool(srvflags.JSONRPCEnableMetrics) {
 		ethmetricsexp.Setup(config.JSONRPC.MetricsAddress)
-	}
-
-	var idxer evertypes.EVMTxIndexer
-	if config.JSONRPC.EnableIndexer {
-		idxDB, err := OpenIndexerDB(home, server.GetAppDBBackend(ctx.Viper))
-		if err != nil {
-			logger.Error("failed to open evm indexer DB", "error", err.Error())
-			return err
-		}
-
-		idxLogger := ctx.Logger.With("indexer", "evm")
-		idxer = indexer.NewKVIndexer(idxDB, idxLogger, clientCtx)
-		indexerService := NewEVMIndexerService(idxer, clientCtx.Client.(rpcclient.Client))
-		indexerService.SetLogger(idxLogger)
-
-		errCh := make(chan error)
-		go func() {
-			if err := indexerService.Start(); err != nil {
-				errCh <- err
-			}
-		}()
-
-		select {
-		case err := <-errCh:
-			return err
-		case <-time.After(types.ServerStartTime): // assume server started successfully
-		}
 	}
 
 	if config.API.Enable || config.JSONRPC.Enable {
@@ -540,6 +510,42 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 	)
 
 	if config.JSONRPC.Enable {
+		// Start EVMTxIndexer service
+		idxDB, err := OpenIndexerDB(home, server.GetAppDBBackend(ctx.Viper))
+		if err != nil {
+			logger.Error("failed to open evm indexer DB", "error", err.Error())
+			return err
+		}
+
+		idxLogger := ctx.Logger.With("indexer", "evm")
+		evmTxIndexer := indexer.NewKVIndexer(idxDB, idxLogger, clientCtx)
+		indexerService := NewEVMIndexerService(evmTxIndexer, clientCtx.Client.(rpcclient.Client))
+		indexerService.SetLogger(idxLogger)
+
+		errCh := make(chan error)
+		go func() {
+			if err := indexerService.Start(); err != nil {
+				errCh <- err
+			}
+		}()
+
+		select {
+		case err := <-errCh:
+			return err
+		case <-time.After(types.ServerStartTime): // assume server started successfully
+		}
+
+		// waiting for indexer indexes blocks
+		for {
+			time.Sleep(1 * time.Second)
+			if evmTxIndexer.IsReady() {
+				break
+			}
+
+			logger.Info("indexer still in progress, keep waiting")
+		}
+
+		// Start Json-RPC server
 		genDoc, err := genDocProvider()
 		if err != nil {
 			return err
@@ -549,7 +555,7 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 
 		tmEndpoint := "/websocket"
 		tmRPCAddr := cfg.RPC.ListenAddress
-		httpSrv, httpSrvDone, err = StartJSONRPC(ctx, clientCtx, tmRPCAddr, tmEndpoint, &config, idxer)
+		httpSrv, httpSrvDone, err = StartJSONRPC(ctx, clientCtx, tmRPCAddr, tmEndpoint, &config, evmTxIndexer)
 		if err != nil {
 			return err
 		}